requirements.txt
tweepy
python-dotenv
requests
instaloader
selenium
pillow
